requirements.txt
anthropic>=0.25.0
Pillow>=10.0.0
lxml>=5.0.0
streamlit>=1.30.0
google-genai>=1.0.0
